Rename constraints.util as constraints.common
