+++ cloup/constraints/_core.py
@@ -9,7 +9,7 @@ from click import Context, Parameter
 
 from cloup._util import check_arg, class_name, make_one_line_repr, make_repr
 from .exceptions import ConstraintViolated, UnsatisfiableConstraint
-from .util import (
+from .common import (
     get_params_whose_value_is_set,
     get_required_params,
     join_param_labels,
