+++ cloup/constraints/conditions.py
@@ -10,7 +10,7 @@ from typing import Any, Generic, TypeVar
 
 from click import Context
 
-from .util import param_label_by_name, param_value_by_name, param_value_is_set
+from .common import param_label_by_name, param_value_by_name, param_value_is_set
 from ._support import ConstraintMixin
 from .._util import make_repr
